+++ benker/grid.py
@@ -37,11 +37,11 @@ You can retrieve the grid cells as follow:
     >>> grid[2, 2] = Cell("blue")
 
     >>> grid[1, 1]
-    <Cell('red', styles={}, nature='body', x=1, y=1, width=1, height=2)>
+    <Cell('red', styles={}, nature=None, x=1, y=1, width=1, height=2)>
     >>> grid[2, 1]
-    <Cell('pink', styles={}, nature='body', x=2, y=1, width=2, height=1)>
+    <Cell('pink', styles={}, nature=None, x=2, y=1, width=2, height=1)>
     >>> grid[2, 2]
-    <Cell('blue', styles={}, nature='body', x=2, y=2, width=1, height=1)>
+    <Cell('blue', styles={}, nature=None, x=2, y=2, width=1, height=1)>
     >>> grid[3, 3]
     Traceback (most recent call last):
         ...
@@ -77,7 +77,7 @@ You can expand the cell size horizontally or vertically:
     >>> grid[2, 2] = Cell("blue")
 
     >>> grid.expand((2, 2), width=1)
-    <Cell('blue', styles={}, nature='body', x=2, y=2, width=2, height=1)>
+    <Cell('blue', styles={}, nature=None, x=2, y=2, width=2, height=1)>
     >>> print(grid)
     +-----------+-----------------------+
     |    red    |   pink                |
@@ -98,7 +98,7 @@ The content of the merged cells is merged too:
     >>> grid[2, 2] = Cell("blue", width=2)
 
     >>> grid.merge((2, 1), (3, 2), content_appender=lambda a, b: "/".join([a, b]))
-    <Cell('pink/blue', styles={}, nature='body', x=2, y=1, width=2, height=2)>
+    <Cell('pink/blue', styles={}, nature=None, x=2, y=1, width=2, height=2)>
     >>> print(grid)
     +-----------+-----------------------+
     |    red    | pink/blue             |